Changes: - Fixed Event dateTime does not save properly
Known Issues:
- When editing Event autofill time is wrong
- Result code is not passed correctly

To be Implemented:
- Decide on essential parameters and add checks for completeness before allowing save
- SnackBar to undo deletion
- Side swipe menu (with pictures) as a hint that left swipe deletes
- Sorting "completed" items into expandable recyclerview items
https://bignerdranch.github.io/expandable-recycler-view/
https://stackoverflow.com/questions/26530685/is-there-an-addheaderview-equivalent-for-recyclerview

+++ app/src/main/java/com/orbital2019/plannerplusplus/model/TaskDao.kt
@@ -19,10 +19,10 @@ interface TaskDao {
     fun update(taskEntity: TaskEntity)
 
     @Query("UPDATE task_table SET isComplete = 1 WHERE id = :id")
-    fun setComplete(id: Int)
+    fun setComplete(id: Long)
 
     @Query("UPDATE task_table SET isComplete = 0 WHERE id = :id")
-    fun setIncomplete(id: Int)
+    fun setIncomplete(id: Long)
 
     @Delete
     fun delete(taskEntity: TaskEntity)
